rename community -> token_pool
