+++ artifacts/checksums.txt
@@ -2,13 +2,13 @@
 f3835b8e9cd79da01f76fca85de427a3f487f31b25ec95b270175f33d27e15e3  brotocol_bbro_minter.wasm
 53277d5b4e3ddd1932c7493c996ae82823e378037d659e1c43ff64131ce6f211  brotocol_bbro_token.wasm
 bf8364b789b0cdafda0d56ba518f9c1ab6b20d468ef72aed7a877cb842e215a4  brotocol_bonding_v1.wasm
-a69197332eb886bdac7cbc562a0db804074931ffdc39f96d5b9fedcf54002506  brotocol_community.wasm
 0b93f03c5f5f051e285486e73e18c4910d98a9f1a4b91b0c21f415f4c2d99980  brotocol_distributor_v1.wasm
 340af29fd9dd1b432dcc92f745b5f6bfb42d49a5a2f57e040048aa8763065a64  brotocol_epoch_manager.wasm
 6b50375f88b9a5890b09db5721954bcde8a3aa3f38186d29ad8c72fc8bd46e67  brotocol_mvp_treasury.wasm
 1c40392405e417a7932224c58590b53976a1f8c12a6964840ea7ae984764b14a  brotocol_oracle.wasm
 018da0a16da9659ca808741c870a082982d7a849dff3a0b460bbe52c5037e231  brotocol_rewards_pool.wasm
 9586c2e54e8df3abf6f40215c338c2e097f11c4b10d974aeed2e554d54abc0f0  brotocol_staking_v1.wasm
+238863b8b667a63594779dc19f2b851a5c5561df40674c1549d0705b9294f59c  brotocol_token_pool.wasm
 0ef15c6ef1416ca9ec4151570ceb4121c487bfb10b11cd7b45455a78e1ecfcf7  brotocol_vesting.wasm
-cbe258db29c5c44d5e684b93a63533e0ccc80f0f60b6764238758d26b0c08675  brotocol_whitelist_sale.wasm
+02b7e2525061cd973bf7cea20e57fbadbbc0009da0ef2b4f39fe075d950390b4  brotocol_whitelist_sale.wasm
 38ce741cb960199a734c8dceed2d9f612e7bf2e5ad430381f973ee816a83cb50  cw20_base.wasm
