+++ contracts/token-pool/schema/config_response.json
