+++ contracts/token-pool/src/commands.rs
