+++ contracts/token-pool/src/error.rs
@@ -2,7 +2,7 @@ use cosmwasm_std::StdError;
 use thiserror::Error;
 
 /// ## Description
-/// This enum describes community pool contract errors
+/// This enum describes token pool contract errors
 #[derive(Error, Debug)]
 pub enum ContractError {
     #[error("{0}")]
