+++ contracts/whitelist-sale/schema/config_response.json
@@ -9,7 +9,7 @@
     "bro_token",
     "owner",
     "rewards_pool_contract",
-    "treasury_contract"
+    "ust_receiver"
   ],
   "properties": {
     "bro_amount_per_nft": {
@@ -40,8 +40,8 @@
       "description": "rewards pool address",
       "type": "string"
     },
-    "treasury_contract": {
-      "description": "treasury contract address",
+    "ust_receiver": {
+      "description": "address for sending received ust",
       "type": "string"
     }
   },
