+++ packages/services/src/lib.rs
@@ -2,13 +2,13 @@ pub mod airdrop;
 pub mod bbro_minter;
 pub mod bonding;
 pub mod common;
-pub mod community;
 pub mod distributor;
 pub mod epoch_manager;
 pub mod oracle;
 pub mod querier;
 pub mod rewards;
 pub mod staking;
+pub mod token_pool;
 pub mod treasury;
 pub mod vesting;
 pub mod whitelist_sale;
