+++ scripts/deploy/lib/config.ts
@@ -17,7 +17,7 @@ export interface Config {
     bbro_token: BbroTokenConfig,
     rewards: RewardsPoolConfig,
     treasury: TreasuryConfig,
-    community: CommunityConfig,
+    token_pool: TokenPoolConfig,
     epoch_manager: EpochManagerConfig,
     stakingv1: StakingV1Config,
     bondingv1: BondingV1Config,
@@ -91,7 +91,7 @@ export interface TreasuryConfig {
     owner: string,
 }
 
-export interface CommunityConfig {
+export interface TokenPoolConfig {
     owner: string,
     bro_token: string,
 }
